서버용2차

+++ springTeamProject/src/main/java/sol/one/VO/ReportVO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class ReportVO {
 
-		private int user_id;
+		private Long user_id;
 		private String re_title;
 		private String re_content;
 		private Date re_time;
