+++ springTeamProject/src/main/java/sol/one/VO/T_tradeVO.java
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class T_tradeVO {
 	  private int product_id;
-	  private int sell_user_id;
-	  private int buyer_user_id;
+	  private Long sell_user_id;
+	  private Long buyer_user_id;
 	  private int pd_status;
 	  private String trade_time;
 }
